pyproject.toml
[project]
name = "dagster-pipeline"
version = "0.1.0"
description = "Dagster ETL pipeline for Google Drive data processing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.1",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
]

[tool.setuptools]
packages = ["dagster_pipeline", "config"]

[dependency-groups]
dagster = [
    "dagster>=1.11.14",
    "dagster-docker>=0.27.14",
    "dagster-postgres>=0.27.14",
]
dev = [
    "black>=25.9.0",
    "pytest>=8.4.2",
]
google-api = [
    "google-api-python-client>=2.184.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
]
